requirements.txt
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
attrs==18.2.0
Babel==2.11.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
bokeh==2.4.3
boto3==1.24.66
botocore==1.27.66
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-completion==0.5.2
click-didyoumean==0.3.0
click-help-colors==0.9.1
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
cloudpickle==2.2.0
colorama==0.4.3
colorcet==3.0.0
cryptography==38.0.1
cycler==0.11.0
Cython==0.29.32
cytoolz==0.11.0
dask==2022.9.1
datashader==0.14.2
datashape==0.5.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distributed==2022.9.1
entrypoints==0.4
executing==1.2.0
facenet-pytorch==2.5.2
fastjsonschema==2.16.2
filelock==3.8.0
fonttools==4.37.1
fsspec==2022.8.2
ftfy==6.1.1
gdown==4.5.1
gitdb==4.0.9
GitPython==3.1.29
gql==3.0.0a6
gradient==2.0.6
gradient-utils==0.5.0
graphql-core==3.1.7
grpcio==1.37.1
grpcio-tools==1.37.1
halo==0.0.31
hdbscan
HeapDict==1.0.1
holoviews==1.15.0
huggingface-hub==0.9.1
idna==3.4
imageio==2.21.2
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
importlib-resources==5.10.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
json5==0.9.10
jsonschema==4.17.0
jupyter-server==1.23.2
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.3
kiwisolver==1.4.4
llvmlite==0.39.1
locket==1.0.0
log-symbols==0.0.14
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==2.21.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
msgpack==1.0.4
multidict==6.0.2
multipledispatch==0.6.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nebula3-database
nest-asyncio==1.5.6
networkx==2.8.6
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.2
numpy==1.23.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
panel==0.13.1
param==1.12.2
paramiko
parso==0.8.3
partd==1.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
progressbar2==4.2.0
prometheus-client==0.9.0
prompt-toolkit==3.0.32
protobuf==3.20.1
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyct==0.4.8
Pygments==2.13.0
PyJWT==2.4.0
pymilvus==1.1.1
pymongo==3.13.0
PyNaCl==1.5.0
pynndescent==0.5.7
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-arango==7.2.0
python-dateutil==2.8.2
python-utils==3.4.5
pytz==2022.2.1
pyviz-comms==2.2.1
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==24.0.1
regex==2022.8.17
requests==2.28.1
requests-toolbelt==0.9.1
s3transfer==0.6.0
sacremoses==0.0.53
scenedetect==0.6.0.3
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
scp
seaborn==0.12.0
Send2Trash==1.8.0
setuptools-scm==7.0.5
shellingham==1.5.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spinners==0.0.24
stack-data==0.6.1
tbb==2021.6.0
tblib==1.7.0
termcolor==2.1.0
terminado==0.17.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2022.8.12
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.12.1
tomli==2.0.1
toolz==0.12.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.18.0
typing_extensions==4.3.0
ujson==5.4.0
umap-learn==0.5.3
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
xarray==2022.6.0
yarl==1.8.1
zict==2.2.0
zipp==3.8.1
weaviate-client==3.11.0
